requirements.txt
opencv-python==4.7.0.72
numpy==1.24.3
pytesseract==0.3.10
Pillow==9.5.0
pdf2image==1.16.3
openai==0.27.2
transformers==4.29.0
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
requests==2.31.0